requirements.txt
Genshi==0.7
requests==2.10.0
vcrpy==1.10.4
beautifulsoup4==4.3.2
python-magic==0.4.12